Show only summary of Pivotal data on dashboard page
The full output of /pivotal is a bit much for the dashboard page; just
show the summary table instead.

I am not sure if this `summary_only` param is the right way to do this,
but it at least frees me from having to set up another endpoint, so
seems like a reasonable trade off for the time-being.

Change-Id: I6d066fd75c1d506c6292c0e35e5fad30aebd2b3d
Reviewed-on: https://gerrit.causes.com/28319
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>