Add breakdown to /pivotal summary table
Show a per-owner breakdown in the summary table, sorted by most numerous
to least, and suppressing the display of ownerless stories.

This is just a proof of concept, and obviously will need to be moved
into a more suitable place (likely a helper) in a subsequent commit. I
am having trouble keeping my eyes open, however, so I wanted to get this
committed even in this ghastly state.

Change-Id: I247008ceafbc555b4cf86c47119ed102b1554b66
Reviewed-on: https://gerrit.causes.com/28322
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>