Use Vary header to address issues with back button in Chrome
Some of our URLs return HTML at times and JSON at others (via an HTTP
Accept header). Chrome helpfully decides to cache those results and
sometimes ends up returning the one you don't want.

For example:

  - starting from the home page, visit the Gerrit tags page; this page
    returns HTML, and then makes a Ajax request for the JSON data to
    populate the page
  - navigate to the home page, then click the back button; at this point
    you'll be looking at a blob of JSON

The fix here, as elsewhere, is to use an HTTP Vary header so that the
browser knows to treat the Ajax request differently from the non-Ajax
one, which means the HTML and JSON are cached separately.

Fixes:

  https://github.com/causes/twigg/issues/16

Included for free:

  - changelog updates

Change-Id: I5d2fe3f6c7a4f1ecfce02ee6b71cd31c52a40413
Reviewed-on: https://gerrit.causes.com/28389
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>