Provide a way to force use of UTF-8 encoding
We've observed some puzzling behavior running Twigg on a CentOS box out
of monit. In the monit environment, we somehow wind up with ASCII-8BIT
as the default encoding, which causes Twigg to choke on UTF-8 encoded
text from places like Git commit log messages and the Pivotal API.

Outside the monit environment, the encoding is UTF-8 and everything is
fine. I am not sure what it is about the environment that makes it
behave differently (I literally copied the entire environment over, one
item at a time, testing as I went, including both obvious candidates
like LANG=en_US.UTF-8, possible suspects like PATH, and left-field ones
like SHELL, TERM and such) and I was unable to get things working inside
monit.

So, here we teach Twigg to take a configuration option that imposes a
sane default encoding. UTF-8 is the default. If you live in a universe
where that doesn't make sense you can override it in your ~/.twiggrc
file.

As we want this setting to be applied to all use of twigg (whether that
be from the command line or from requiring the gem in another context
like `irb` or a different project), we've rejigged the config process
ever so slightly to make sure that at boot time (ie. as soon as anybody
does `require 'twigg'`) we set up a sane encoding.

Fixes:

  https://github.com/causes/twigg/issues/15

Change-Id: I56185b795a51cab7e69ca5cf2d8e095a643ef565
Reviewed-on: https://gerrit.causes.com/28425
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>