Loosen dependencies on gems which depend on main twigg gem
In order to make installation and upgrades easier, this commit makes the
various subordinate gems in the Twigg project depend more loosely on the
main twigg gem.

Before this commit:

  - twigg: no hard dependencies; soft dependency on subordinate gems for
    optional features
  - twigg-*: explicit hard dependency on latest version of main twigg
    gem

And additionally:

  - twigg-app: soft dependency on other gems for optional features

Depending on the latest version of the main twigg gem can create a
number of problems during updates. For example, say I want to update
"twigg-app" and "twigg" at the same time. I install the new version of
"twigg" and "twigg-app" is happy about it, but the other gems that
depend on "twigg" and which haven't been upgraded will want the older
version of it, rendering the server at best degraded (without the
optional functionality) and at worst inoperable (unable to boot).

With this commit:

  - twigg: no hard dependencies; soft "dependency" on subordinate gems
    for optional features (ie. same as before)
  - twigg-*: hard dependency on recent version of main twigg gem (ie.
    this is changed)
  - twigg-app: soft dependency on other gems for optional features (ie.
    same as before)

This doesn't enable all possible workflows, but it enables at least
these ones:

  - Update main twigg gem; all the other gems will accept the new
    version
  - Update any subordinate gem in isolation; it will continue to depend
    on "~> Twigg::VERSION", which in this case is unchanged, so
    everything keeps working as is
  - Update main twigg gem and one other gem at the same time; again,
    the other subordinate gems will happily use the new gem

There are edge cases that won't work in this model, but if we ever run
into them we can switch to something more drastic (or pedestrian) like
manually bumping the dependency version numbers (ie. micro-managing).

Change-Id: I45cd769055660ef0c8acb36a6bcf6316633810bf
Reviewed-on: https://gerrit.causes.com/28427
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>